refactor(flutter): change all references from `flutter-styles` to `flutter`
BREAKING CHANGE: Now `facio_components` Flutter package will be named `facio_design_system`. Folder `flutter-styles` was renamed to `flutter`
