+++ .github/scripts/bump-pubspec.sh
@@ -5,10 +5,10 @@ set -e
 git status
 
 # Prepare Flutter pubspec file to commit
-git add ./packages/flutter-styles/pubspec.yaml
+git add ./packages/flutter/pubspec.yaml
 
 # Commit pubspec file
-git commit --allow-empty -m "chore(flutter-styles): bump package version [skip ci]"
+git commit --allow-empty -m "chore(flutter): bump package version [skip ci]"
 
 # Push up-to-date pubspec file
 git push origin
