+++ .github/workflows/publish.yml
@@ -43,8 +43,8 @@ jobs:
           cache-name: cache-pubspec-file
         with:
           path: |
-            packages/flutter-styles/pubspec.yaml
-          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('packages/flutter-styles/pubspec.yaml') }}
+            packages/flutter/pubspec.yaml
+          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('packages/flutter/pubspec.yaml') }}
           restore-keys: |
             ${{ runner.os }}-build-${{ env.cache-name }}-
 
@@ -118,7 +118,7 @@ jobs:
     needs: build
     runs-on: ubuntu-latest
     env:
-      working-directory: ./packages/flutter-styles
+      working-directory: ./packages/flutter
     steps:
       - name: Checkout repository
         uses: actions/checkout@v2
@@ -130,8 +130,8 @@ jobs:
           cache-name: cache-pubspec-file
         with:
           path: |
-            packages/flutter-styles/pubspec.yaml
-          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('packages/flutter-styles/pubspec.yaml') }}
+            packages/flutter/pubspec.yaml
+          key: ${{ runner.os }}-build-${{ env.cache-name }}-${{ hashFiles('packages/flutter/pubspec.yaml') }}
           restore-keys: |
             ${{ runner.os }}-build-${{ env.cache-name }}-
       - name: Install Flutter SDK
