+++ packages/flutter/example/pubspec.yaml
@@ -10,7 +10,7 @@ environment:
   sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
-  components:
+  facio_design_system:
     path: ..
   flutter:
     sdk: flutter
