+++ packages/flutter/lib/components.dart
@@ -1,2 +1,3 @@
 export 'src/animations.dart';
 export 'src/components.dart';
+export 'styles.dart';
