+++ packages/flutter/lib/src/components.dart
@@ -7,17 +7,12 @@ export 'components/legacy/button.dart';
 export 'components/legacy/canvas.dart';
 export 'components/legacy/card.dart';
 export 'components/legacy/chips.dart';
-export 'components/legacy/debouncer.dart';
-export 'components/legacy/emoji_text.dart';
-export 'components/legacy/hide_keyboard.dart';
 export 'components/legacy/legacy_theme.dart';
-export 'components/legacy/pin_code.dart';
 export 'components/legacy/swipe_button.dart';
 export 'components/legacy/text_bullet.dart';
 export 'components/legacy/video_thumb.dart';
 export 'components/name_field.dart';
 export 'components/primary_button_contained.dart';
 export 'components/progress_widget.dart';
-export 'components/scroll_view.dart';
 export 'components/tertiary_button_link.dart';
 export 'components/tertiary_button_outlined.dart';
