+++ packages/flutter/lib/src/components/name_field.dart
@@ -1,4 +1,4 @@
-import 'package:facio_styles/facio_styles.dart';
+import 'package:facio_design_system/components.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
