+++ packages/flutter/pubspec.yaml
@@ -1,6 +1,6 @@
 ---
-name: facio_components
-description: Facio Design System UI components of Facio applications built on Flutter.
+name: facio_design_system
+description: Facio Design System style tokens and UI components for Facio applications built on Flutter.
 version: 0.30.3
 homepage: https://github.com/FacioCode/design/tree/main/packages/flutter#readme
 repository: https://github.com/FacioCode/design
