+++ packages/style-dictionary/package.json
@@ -34,7 +34,7 @@
   },
   "scripts": {
     "build": "npm run clean && npm run compile",
-    "clean": "style-dictionary clean --config ./src/index.js && rm -rf ./dist && rm -rf ../flutter-styles/lib",
+    "clean": "style-dictionary clean --config ./src/index.js && rm -rf ./dist && rm -rf ../flutter/lib/src/generated",
     "compile": "style-dictionary build --config ./src/index.js",
     "test": "echo \"Todo: implement tests\" && exit 0"
   }
