+++ packages/style-dictionary/src/templates/flutter/component_styles
@@ -1,4 +1,7 @@
-part of <%= packageName %>;
+import 'dart:ui';
+import 'package:flutter/material.dart';
+import 'package:flutter/painting.dart';
+
 <%
 var getValuePrefix = ({attributes}) => {
   const { subitem } = attributes;
