+++ packages/style-dictionary/src/templates/flutter/library_class
@@ -3,7 +3,9 @@ var props = _.sortBy(allProperties, item => item.attributes.category + item.name
 var getNamePrefix = (category) => category === "color" ? "Color " : "";
 var getValuePrefix = (category) => category === "color" ? "const ": "";
 
-%>part of <%= this.packageName %>;
+%>import 'dart:ui';
+import 'package:flutter/material.dart';
+import 'package:flutter/painting.dart';
 
 class <%= this.className %> {
   <%= this.className %>._();
